backend/requirements.txt
django
django-cors-headers
djangorestframework_simplejwt
Pillow
python-socketio